requirements.txt
streamlit==1.28.1
pdfplumber==0.9.0
python-docx==0.8.11
pytesseract==0.3.10
pillow==10.0.0
scikit-learn==1.3.2
numpy==1.24.3
pandas==2.0.3
sentence-transformers==2.2.2
torch==2.0.1
spacy==3.6.1
nltk==3.8.1
xgboost==2.0.0
matplotlib==3.8.1
seaborn==0.13.0
plotly==5.17.0
joblib==1.3.2
python-pptx==0.6.21
